Bugfix: Use WindowsFocusListener not FocusListener
